requirements/base.txt
asgiref==3.6.0
Django==4.1.4
django-environ==0.9.0
djangorestframework==3.14.0
numpy==1.24.0
pandas==1.5.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
